feat: Prepare updating of colors when phase changes
The background and the timer background should have the same color,
only differing a bit in brightness. However, to control the background
color, we need the background to be contained within the Yew component.
Hence, we cover the page with an infinitely large `div` that we can
manipulate.

Some style changes were made to correct the visual aspect after this
addition.